+++ app/src/main/java/jp/co/tbdeveloper/warikanapp/feature_roulette/domain/repository/RouletteRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface RouletteRepository {
     fun getRoulettes(): Flow<List<RouletteEntity>>
-    suspend fun getRouletteById(id: Int): RouletteEntity?
+    suspend fun getRouletteById(id: Long): RouletteEntity?
     suspend fun insertRoulette(roulette: RouletteEntity): Long
     suspend fun deleteRoulette(roulette: RouletteEntity)
 }
